+++ keyboards/boardsource/microdox/info.json
@@ -1,54 +1,70 @@
 {
-    "keyboard_name": "microdox",
-    "url": "https://boardsource.xyz/store/5f2e7e4a2902de7151494f92",
-    "maintainer": "boardsource",
-    "layouts": {
-        "LAYOUT_split_3x5_3": {
-            "layout": [
-                { "label": "Q", "x": 0, "y": 0.3 },
-                { "label": "W", "x": 1, "y": 0.1 },
-                { "label": "E", "x": 2, "y": 0 },
-                { "label": "R", "x": 3, "y": 0.1 },
-                { "label": "T", "x": 4, "y": 0.2 },
-
-                { "label": "Y", "x": 8, "y": 0.2 },
-                { "label": "U", "x": 9, "y": 0.1 },
-                { "label": "I", "x": 10, "y": 0 },
-                { "label": "O", "x": 11, "y": 0.1 },
-                { "label": "P", "x": 12, "y": 0.3 },
-
-                { "label": "A", "x": 0, "y": 1.3 },
-                { "label": "S", "x": 1, "y": 1.1 },
-                { "label": "D", "x": 2, "y": 1 },
-                { "label": "F", "x": 3, "y": 1.1 },
-                { "label": "G", "x": 4, "y": 1.2 },
-
-                { "label": "H", "x": 8, "y": 1.2 },
-                { "label": "J", "x": 9, "y": 1.1 },
-                { "label": "K", "x": 10, "y": 1 },
-                { "label": "L", "x": 11, "y": 1.1 },
-                { "label": ";", "x": 12, "y": 1.3 },
-
-                { "label": "Z", "x": 0, "y": 2.3 },
-                { "label": "X", "x": 1, "y": 2.1 },
-                { "label": "C", "x": 2, "y": 2 },
-                { "label": "V", "x": 3, "y": 2.1 },
-                { "label": "B", "x": 4, "y": 2.2 },
-
-                { "label": "N", "x": 8, "y": 2.2 },
-                { "label": "M", "x": 9, "y": 2.1 },
-                { "label": ",", "x": 10, "y": 2 },
-                { "label": ".", "x": 11, "y": 2.1 },
-                { "label": "/", "x": 12, "y": 2.3 },
-
-                { "label": "GUI / KC_HANJ", "x": 3, "y": 3.7 },
-                { "label": "Lower", "x": 4, "y": 3.7 },
-                { "label": "Space", "x": 5, "y": 3.2 },
-
-                { "label": "Enter", "x": 7, "y": 3.2 },
-                { "label": "Raise", "x": 8, "y": 3.7 },
-                { "label": "Alt / KC_HAEN", "x": 9, "y": 3.7 }
-            ]
-        }
+  "manufacturer": "Boardsource",
+  "keyboard_name": "microdox",
+  "maintainer": "waffle87",
+  "development_board": "promicro",
+  "diode_direction": "COL2ROW",
+  "build": {
+    "lto": true
+  },
+  "features": {
+    "bootmagic": true,
+    "extrakey": true,
+    "mousekey": true,
+    "nkro": true,
+    "oled": true
+  },
+  "url": "https://boardsource.xyz/store/5f2e7e4a2902de7151494f92",
+  "usb": {
+    "pid": "0x0412",
+    "vid": "0xF7E0"
+  },
+  "split": {
+    "enabled": true
+  },
+  "community_layouts": [
+    "split_3x5_3"
+  ],
+  "layouts": {
+    "LAYOUT_split_3x5_3": {
+      "layout": [
+        { "matrix": [0, 0], "x": 0, "y": 0 },
+        { "matrix": [0, 1], "x": 1, "y": 0 },
+        { "matrix": [0, 2], "x": 2, "y": 0 },
+        { "matrix": [0, 3], "x": 3, "y": 0 },
+        { "matrix": [0, 4], "x": 4, "y": 0 },
+        { "matrix": [4, 4], "x": 5, "y": 0 },
+        { "matrix": [4, 3], "x": 6, "y": 0 },
+        { "matrix": [4, 2], "x": 7, "y": 0 },
+        { "matrix": [4, 1], "x": 8, "y": 0 },
+        { "matrix": [4, 0], "x": 9, "y": 0 },
+        { "matrix": [1, 0], "x": 10, "y": 0 },
+        { "matrix": [1, 1], "x": 11, "y": 0 },
+        { "matrix": [1, 2], "x": 12, "y": 0 },
+        { "matrix": [1, 3], "x": 13, "y": 0 },
+        { "matrix": [1, 4], "x": 14, "y": 0 },
+        { "matrix": [5, 4], "x": 15, "y": 0 },
+        { "matrix": [5, 3], "x": 16, "y": 0 },
+        { "matrix": [5, 2], "x": 17, "y": 0 },
+        { "matrix": [5, 1], "x": 18, "y": 0 },
+        { "matrix": [5, 0], "x": 19, "y": 0 },
+        { "matrix": [2, 0], "x": 20, "y": 0 },
+        { "matrix": [2, 1], "x": 21, "y": 0 },
+        { "matrix": [2, 2], "x": 22, "y": 0 },
+        { "matrix": [2, 3], "x": 23, "y": 0 },
+        { "matrix": [2, 4], "x": 24, "y": 0 },
+        { "matrix": [6, 4], "x": 25, "y": 0 },
+        { "matrix": [6, 3], "x": 26, "y": 0 },
+        { "matrix": [6, 2], "x": 27, "y": 0 },
+        { "matrix": [6, 1], "x": 28, "y": 0 },
+        { "matrix": [6, 0], "x": 29, "y": 0 },
+        { "matrix": [3, 2], "x": 30, "y": 0 },
+        { "matrix": [3, 3], "x": 31, "y": 0 },
+        { "matrix": [3, 4], "x": 32, "y": 0 },
+        { "matrix": [7, 4], "x": 33, "y": 0 },
+        { "matrix": [7, 3], "x": 34, "y": 0 },
+        { "matrix": [7, 2], "x": 35, "y": 0 }
+      ]
     }
+  }
 }
