--- keyboards/boardsource/microdox/keymaps/cole/rules.mk
@@ -1,1 +1,0 @@
-OLED_ENABLE = yes
