+++ keyboards/boardsource/microdox/keymaps/via/rules.mk
@@ -1,3 +1,1 @@
-OLED_ENABLE = yes
 VIA_ENABLE = yes
-LTO_ENABLE = yes
