+++ keyboards/boardsource/microdox/microdox.c
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 Cole Smith <cole@boadsource.xyz>
+Copyright 2022 Cole Smith <cole@boadsource.xyz>
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -15,9 +15,11 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "microdox.h"
-
 #ifdef OLED_ENABLE
+#include "oled_driver.h"
+#include "bitwise.h"
+#include "action_layer.h"
+
 oled_rotation_t oled_init_kb(oled_rotation_t rotation) {
   if (is_keyboard_master())
     return OLED_ROTATION_180;
@@ -26,14 +28,12 @@ oled_rotation_t oled_init_kb(oled_rotation_t rotation) {
 
 static void render_logo(void) {
   static const char PROGMEM qmk_logo[] = {
-    0x80,0x81,0x82,0x83,0x84,0x85,0x86,0x87,0x88,0x89,0x8a,0x8b,0x8c,0x8d,0x8e,0x8f,0x90,0x91,0x92,0x93,0x94,
-    0xa0,0xa1,0xa2,0xa3,0xa4,0xa5,0xa6,0xa7,0xa8,0xa9,0xaa,0xab,0xac,0xad,0xae,0xaf,0xb0,0xb1,0xb2,0xb3,0xb4,
-    0xc0,0xc1,0xc2,0xc3,0xc4,0xc5,0xc6,0xc7,0xc8,0xc9,0xca,0xcb,0xcc,0xcd,0xce,0xcf,0xd0,0xd1,0xd2,0xd3,0xd4,
-    0
+    0x80, 0x81, 0x82, 0x83, 0x84, 0x85, 0x86, 0x87, 0x88, 0x89, 0x8A, 0x8B, 0x8C, 0x8D, 0x8E, 0x8F, 0x90, 0x91, 0x92, 0x93, 0x94,
+    0xA0, 0xA1, 0xA2, 0xA3, 0xA4, 0xA5, 0xA6, 0xA7, 0xA8, 0xA9, 0xAA, 0xAB, 0xAC, 0xAD, 0xAE, 0xAF, 0xB0, 0xB1, 0xB2, 0xB3, 0xB4,
+    0xC0, 0xC1, 0xC2, 0xC3, 0xC4, 0xC5, 0xC6, 0xC7, 0xC8, 0xC9, 0xCA, 0xCB, 0xCC, 0xCD, 0xCE, 0xCF, 0xD0, 0xD1, 0xD2, 0xD3, 0xD4, 0x00
   };
   oled_write_P(qmk_logo, false);
 }
-
 static void render_status(void) {
    switch (get_highest_layer(layer_state)) {
     case 0:
@@ -72,5 +72,26 @@ bool oled_task_kb(void) {
     oled_scroll_left();
   }
   return false;
+}
+#endif
+
+#ifdef ENCODER_ENABLE
+#include "encoder.h"
+bool encoder_update_kb(uint8_t index, bool clockwise) {
+  if (!encoder_update_user(index, clockwise)) { return false; }
+  if (index == 0) {
+    if (clockwise) {
+      tap_code_delay(KC_VOLU, 10);
+    } else {
+      tap_code_delay(KC_VOLD, 10);  
+    }
+  } else {
+    if (clockwise) {
+      tap_code(KC_MNXT);
+    } else {
+      tap_code(KC_MPRV);
+    }
+  }
+  return true;
 }
 #endif
