+++ keyboards/boardsource/microdox/readme.md
@@ -8,8 +8,9 @@
 
 Make example for this keyboard (after setting up your build environment):
 
+    make boardsource/microdox/v1:default
     make boardsource/microdox/v2:default
 
-Remove `v2` from above command if you purchased your PCBs prior to April 2022
+Compile `v1` firmware if you purchased your PCBs prior to April 2022.
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
