+++ keyboards/boardsource/microdox/v1/info.json
@@ -1,0 +1,25 @@
+{
+  "features": {
+    "rgblight": true
+  },
+    "matrix_pins": {
+    "cols": ["F4", "F5", "F6", "F7", "B1"],
+    "rows": ["B2", "B6", "B4", "B5"]
+  },
+  "usb": {
+    "device_version": "1.0.0"
+  },
+  "split": {
+    "soft_serial_pin": "D2"
+  },
+    "rgblight": {
+    "pin": "D3",
+    "sleep": true,
+    "led_count": 12,
+    "split_count": [6, 6],
+    "max_brightness": 150,
+    "animations": {
+      "all": true
+    }
+  }
+}
