+++ keyboards/boardsource/microdox/v1/rules.mk
@@ -1,0 +1,1 @@
+# This file intentionally left blank
