+++ keyboards/boardsource/microdox/v2/info.json
@@ -1,0 +1,82 @@
+{
+  "features": {
+    "rgb_matrix": true
+  },
+  "matrix_pins": {
+    "cols": ["D4", "D7", "B3", "F7", "F6"],
+    "rows": ["F4", "D2", "C6", "B1"]
+  },
+  "usb": {
+    "device_version": "2.0.0"
+  },
+  "encoder": {
+    "enabled": true,
+    "rotary": [
+      { "pin_a": "E6", "pin_b": "B2"}
+    ]
+  },
+  "split": {
+    "soft_serial_pin": "D3",
+    "matrix_pins": {
+      "right": {
+        "cols": ["F4", "B1", "D7", "C6", "B3"],
+        "rows": ["F5", "F7", "F6", "E6"]
+      }
+    },
+    "encoder": {
+      "right": {
+        "rotary": [
+          { "pin_a": "B6", "pin_b": "B2" }
+        ]
+      }
+    }
+  },
+  "rgb_matrix": {
+    "layout": [
+      { "flags": 2, "x": 12, "y": 11 },
+      { "flags": 2, "x": 86, "y": 11 },
+      { "flags": 2, "x": 86, "y": 53 },
+      { "flags": 2, "x": 0, "y": 64 },
+      { "flags": 2, "x": 211, "y": 11 },
+      { "flags": 2, "x": 136, "y": 11 },
+      { "flags": 2, "x": 136, "y": 53 },
+      { "flags": 2, "x": 224, "y": 64 },
+      { "flags": 4, "matrix": [0, 0], "x": 0,  "y": 0 },
+      { "flags": 4, "matrix": [0, 1], "x": 24,  "y": 0 },
+      { "flags": 4, "matrix": [0, 2], "x": 49,  "y": 0 },
+      { "flags": 4, "matrix": [0, 3], "x": 74,  "y": 0 },
+      { "flags": 4, "matrix": [0, 4], "x": 99,  "y": 0 },
+      { "flags": 4, "matrix": [4, 4], "x": 124,  "y": 0 },
+      { "flags": 4, "matrix": [4, 3], "x": 149,  "y": 0 },
+      { "flags": 4, "matrix": [4, 2], "x": 174,  "y": 0 },
+      { "flags": 4, "matrix": [4, 1], "x": 199,  "y": 0 },
+      { "flags": 4, "matrix": [4, 0], "x": 224,  "y": 0 },
+      { "flags": 4, "matrix": [1, 0], "x": 0,  "y": 21 },
+      { "flags": 4, "matrix": [1, 1], "x": 24,  "y": 21 },
+      { "flags": 4, "matrix": [1, 2], "x": 49,  "y": 21 },
+      { "flags": 4, "matrix": [1, 3], "x": 74,  "y": 21 },
+      { "flags": 4, "matrix": [1, 4], "x": 99,  "y": 21 },
+      { "flags": 4, "matrix": [5, 4], "x": 124,  "y": 21 },
+      { "flags": 4, "matrix": [5, 3], "x": 149,  "y": 21 },
+      { "flags": 4, "matrix": [5, 2], "x": 174,  "y": 21 },
+      { "flags": 4, "matrix": [5, 1], "x": 199,  "y": 21 },
+      { "flags": 4, "matrix": [5, 0], "x": 224,  "y": 21 },
+      { "flags": 4, "matrix": [2, 0], "x": 0,  "y": 42 },
+      { "flags": 4, "matrix": [2, 1], "x": 24,  "y": 42 },
+      { "flags": 4, "matrix": [2, 2], "x": 49,  "y": 42 },
+      { "flags": 4, "matrix": [2, 3], "x": 74,  "y": 42 },
+      { "flags": 4, "matrix": [2, 4], "x": 99,  "y": 42 },
+      { "flags": 4, "matrix": [6, 4], "x": 124,  "y": 42 },
+      { "flags": 4, "matrix": [6, 3], "x": 149,  "y": 42 },
+      { "flags": 4, "matrix": [6, 2], "x": 174,  "y": 42 },
+      { "flags": 4, "matrix": [6, 1], "x": 199,  "y": 42 },
+      { "flags": 4, "matrix": [6, 0], "x": 224,  "y": 42 },
+      { "flags": 1, "matrix": [3, 2], "x": 49,  "y": 64 },
+      { "flags": 1, "matrix": [3, 3], "x": 74,  "y": 64 },
+      { "flags": 1, "matrix": [3, 4], "x": 99,  "y": 64 },
+      { "flags": 1, "matrix": [7, 4], "x": 124,  "y": 64 },
+      { "flags": 1, "matrix": [7, 3], "x": 149,  "y": 64 },
+      { "flags": 1, "matrix": [7, 2], "x": 174,  "y": 64 }
+    ]
+  }
+}
