+++ keyboards/boardsource/microdox/v2/rules.mk
@@ -1,8 +1,1 @@
-RGB_MATRIX_ENABLE = yes
 RGB_MATRIX_DRIVER = WS2812
-RGB_MATRIX_SUPPORTED = yes
-RGBLIGHT_ENABLE = no
-OLED_ENABLE = yes
-OLED_DRIVER = SSD1306
-ENCODER_ENABLE = yes
-LTO_ENABLE = yes
